fix(comments) Do not throw if comment has no footer
A misformatted comment can screw up the the whole page markup so the
footer can not be found in the regular path.
The patch checks if footer exists before accessing to it.

The comments after the screwed comment can be outside of the comments
container, they still will not be found.

Github issue #23