fix(storage) Additional check for storage working
The extension wants to be ready to use sync storage whenever it's
available. To choose the storage was pretty simple, if
chrome.storage.sync exists, use it, otherwise use chrome.storage.local.

In Firefox 52 chrome.storage.sync has been added to the API, but it's
disabled in about:config. It notifies in chrome.runtime.lastError that
something is wrong, but the API still exposed. The assumption, that the
API will be exposed only if it's ready to use is too optimistic, so
added some additional check, which made the getStorageArea function more
complex and async. It also returns a promise now.

Github issue #26