feat(validation) Add html format validitaion
Test if the comment the user tries to send has any unclosed tag.
The test is very dumb: create a div set the comment text as innerHTML.
Then get the innerHTML of the same div, if the result is not the same
what were set, there must be some unclosed tag, because the browser will
add the closing tags automtically.