fix(validator) Prevent form submit
The FormSubmitEvent.preventDefault was called in a callback of the
validator then branch. Promise callbacks are scheduled in a different
queue, so it may happen that the preventDefault would be called after
the browser submits the form.
Solution is to call the preventDefault in the beginning of the form
submit handler.

To allow to submit the form, a novalidate class on the form will abort
the submit event handler.