fix(validator) Submit form on first click
Chrome and Firefox behaves differently when it's about to handle
promises.
For unkown reason by myself, a timeout is needed before simulating the
click event of a valid form button to really submit the form.
Probably the form's default action still prevented by preventDefault and
enabled when the event loop starts the next task queue, which is
probably after all promises resolved.