refactor(content-block) Remove hardcoded twitter url
Remove twitter widget script url from the source, instead use script url
queried from the page.

That hopeufully fixes issue on AMO, where they blocked the addon version
2.5.0 because it's loading an external script.
That script would be loaded anyway to the page, we just block it. And
add it again to the page to reenable it.